Move gaussian_kernel to convolution.py
